requirements.txt
pip>=21.0
termcolor>=1.1.0
langchain==0.0.275
python-dotenv
pandas
openai==0.28
scikit-learn
tqdm